requirements.txt

# Google Cloud Functions / runtime helpers
functions-framework
google-cloud-firestore
flask
flask-cors
gunicorn
google-auth

# Developer tools
black
flake8
pytest-cov

# Core dependencies
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
jupyterlab>=4.0.0
joblib>=1.3.0

# FastAPI and web server
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0

# Authentication dependencies
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0
pydantic>=2.0.0
